SHR-3: Refactoring of PnR methods into service to better allow for rollback
