+++ api/src/main/java/org/openmrs/module/xdsbrepository/XDSbService.java
@@ -3,6 +3,8 @@ package org.openmrs.module.xdsbrepository;
 import java.net.MalformedURLException;
 import java.util.Map;
 
+import org.dcm4chee.xds2.common.exception.XDSException;
+import org.dcm4chee.xds2.infoset.ihe.ProvideAndRegisterDocumentSetRequestType;
 import org.dcm4chee.xds2.infoset.rim.RegistryResponseType;
 import org.dcm4chee.xds2.infoset.rim.SubmitObjectsRequest;
 import org.openmrs.api.OpenmrsService;
@@ -22,8 +24,8 @@ public interface XDSbService extends OpenmrsService {
      * @throws Exception
      * @throws MalformedURLException
      */
-    public RegistryResponseType registerDocument(String documentUniqueId, Class<? extends ContentHandler> contentHandler, SubmitObjectsRequest submitObjectRequest) throws Exception;
-	
+    RegistryResponseType registerDocument(String documentUniqueId, Class<? extends ContentHandler> contentHandler, SubmitObjectsRequest submitObjectRequest) throws XDSException;
+
 	/**
 	 * Registers documents with the configured XDS.b registry and stores a mapping
 	 * from documentUniqueId to the supplied ContentHandler class for each document.
@@ -33,7 +35,7 @@ public interface XDSbService extends OpenmrsService {
 	 * @throws Exception 
 	 * @throws MalformedURLException 
 	 */
-	public RegistryResponseType registerDocuments(Map<String, Class<? extends ContentHandler>> contentHandlers, SubmitObjectsRequest submitObjectRequest) throws Exception;
+	RegistryResponseType registerDocuments(Map<String, Class<? extends ContentHandler>> contentHandlers, SubmitObjectsRequest submitObjectRequest) throws XDSException;
 	
 	/**
 	 * Fetches the content handler class that can retrieve the given documentUniqueId.
@@ -42,6 +44,10 @@ public interface XDSbService extends OpenmrsService {
 	 * @return The class of the content handler that can retrieve this document.
 	 * @throws ClassNotFoundException if the found class cannot be loaded
 	 */
-	public Class<? extends ContentHandler> getDocumentHandlerClass(String documentUniqueId) throws ClassNotFoundException;
+	Class<? extends ContentHandler> getDocumentHandlerClass(String documentUniqueId) throws ClassNotFoundException;
 
+	/**
+	 * Processes an XDS.b Provide and register document request
+	 */
+	RegistryResponseType provideAndRegisterDocumentSetB(ProvideAndRegisterDocumentSetRequestType request) throws XDSException;
 }
