+++ api/src/main/java/org/openmrs/module/xdsbrepository/exceptions/UnsupportedGenderException.java
@@ -1,4 +1,5 @@
-package org.openmrs.module.xdsbrepository.ihe.iti.actors.impl.exceptions;
+package org.openmrs.module.xdsbrepository.exceptions;
+
 
 public class UnsupportedGenderException extends XdsRepositoryException {
