+++ api/src/main/java/org/openmrs/module/xdsbrepository/exceptions/XdsRepositoryException.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.xdsbrepository.ihe.iti.actors.impl.exceptions;
+package org.openmrs.module.xdsbrepository.exceptions;
 
 /**
  * XdsRepositoryException base exception for all repository exceptions
